+++ examples/get_private_key.py
@@ -1,6 +1,6 @@
 import sys
 
-from gufo.acme.client import ACMEClient
+from gufo.acme.clients.base import ACMEClient
 
 
 def main(path: str) -> None: