+++ src/gufo/acme/clients/__init__.py
@@ -1,0 +1,13 @@
+# ---------------------------------------------------------------------
+# Gufo ACME: Various clients implementation.
+# ---------------------------------------------------------------------
+# Copyright (C) 2023, Gufo Labs
+# ---------------------------------------------------------------------
+
+"""
+Various ACME clients implementations.
+
+The following clients are provided out-of-the-box:
+
+* [base
+"""